+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Please use [this](https://docs.gitlab.com/ee/development/changelog.html) document as guidelines to keep a changelog.
 
+## [unreleased]
+
+### Removed
+
+* bcd: Removed `bcd` datasource and config section.
+
 ## 4.2.6 - 2022-02-25
 
 ### Fixed